LayoutTests/platform/wpe/fast/scrolling/sync-scroll-overscroll-behavior-unscrollable-element-expected.txt


FAIL Unscrollable element horizontal scroll test with overscroll-behavior: none. assert_equals: Propagate horizontal scroll expected true but got false
PASS Unscrollable element vertical scroll test with overscroll-behavior: none.
